feat: Adicionado sistema de Authentication e Authorization

+++ Estoque.Backend/Models/Product.cs
@@ -6,7 +6,7 @@
 
         public string image { get; set; }
         public string name { get; set; }
-        public decimal price { get; set; }
+        public double price { get; set; }
 
         public int stock { get; set; }
     }
@@ -15,7 +15,7 @@
     {
         public string image { get; set; }
         public string name { get; set; }
-        public decimal price { get; set; }
+        public double price { get; set; }
 
         public int stock { get; set; }
